+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ## Fixed
 - Fix modify userinfo service to include scope in response [mabhi](https://github.com/mabhi)
 
+## Changed
+- Upgraded Ory Kratos to v0.10.1 [akshay196](https://github.com/akshay196)
+
 ## [0.1.8] - 2022-11-25
 
 ## Added