+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ## Fixed
 - Fix modify userinfo service to include scope in response [mabhi](https://github.com/mabhi)
+- Kubectl commands work even after deleting an imported cluster from [niravparikh05](https://github.com/niravparikh05)
 
 ## [0.1.8] - 2022-11-25