+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## [0.1.9] - 2022-12-29
+
 ## Added
 - Enhance: record user.login event via kratos hooks [mabhi](https://github.com/mabhi)
 
@@ -67,7 +69,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/paralus/paralus/compare/v0.1.8...HEAD
+[Unreleased]: https://github.com/paralus/paralus/compare/v0.1.9...HEAD
+[0.1.9]: https://github.com/paralus/paralus/compare/v0.1.8...v0.1.9
 [0.1.8]: https://github.com/paralus/paralus/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/paralus/paralus/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/paralus/paralus/compare/v0.1.5...v0.1.6